+++ app/Http/Controllers/OwnerController.php
@@ -32,13 +32,19 @@ class OwnerController extends Controller
         return redirect()->route('owners.index');
     }
 
-    public function edit(Owner $owner)
+    public function edit(Request $request, Owner $owner)
     {
+        if (! $request->user()->can('editOwner', $owner) ){
+            return redirect()->route('owners.index');
+        }
         return view('owners.edit', ['owner' => $owner]);
     }
 
     public function update(Request $request, Owner $owner)
     {
+        if (! $request->user()->can('editOwner', $owner) ){
+            return redirect()->route('owners.index');
+        }
         $owner->name = $request->name;
         $owner->surname = $request->surname;
         $owner->phone = $request->phone;
@@ -49,8 +55,11 @@ class OwnerController extends Controller
         return redirect()->route('owners.index');
     }
 
-    public function destroy(Owner $owner)
+    public function destroy(Request $request, Owner $owner)
     {
+        if (! $request->user()->can('deleteOwner', $owner) ){
+            return redirect()->route('owners.index');
+        }
         $owner->delete();
         return redirect()->route('owners.index');
     }
